test_project/requirements.txt
Django>=2.2.3,<2.2.5
djangorestframework>=3.9.0,<3.10.3
drf-rw-serializers==1.0.2
twilio~=6.38.0
